+++ Changelog.md
@@ -1,3 +1,27 @@
+# Rework 3.0: Langraph
+
+## Limit
+
+- Loop call
+- No iteration
+
+## Further
+
+- Compilation Structure for submodule of a complex module.
+- Another LLM for Stating the current module/submodule.
+
+## 2.0.0
+
+- Speed up with Langraph
+
+- No iteration for framework.
+  - Only for looping.
+- Memory
+
+- Init: Chat with initial code.
+- Single module compilation only.
+- Prompt Template for the Code Fixer LLM.
+  - Json/dict format for the output
 # Purpose
 
 - Improve RTLLM
@@ -21,6 +45,7 @@
 - Templating Makefile: run command of verilator [x]
 - LLM Generation History [x].
 - Click alternative to ArgumentParser.
+- Tuple alternative for speeding!
 
 # 0.3.0